Eth beta test

+++ Multy/ViewControllers and Presenters/Donation/SendDonation/DonationSendPresenter.swift
@@ -32,7 +32,7 @@ class DonationSendPresenter: NSObject, CustomFeeRateProtocol, SendWalletProtocol
             selectedIndexOfSpeed = 2
         }
         let cell = self.mainVC?.tableView.cellForRow(at: [0, selectedIndexOfSpeed!]) as! CustomTrasanctionFeeTableViewCell
-        cell.value = (firstValue)!
+        cell.value = UInt64(firstValue!)
         cell.setupUIFor(gasPrice: nil, gasLimit: nil)
         self.mainVC?.isTransactionSelected = true
         self.customFee = UInt64(firstValue!)
